- Fixed an issue where the page to show all Flow Outs would just crash when trying to open it by replacing it - Switched from Uranium UI datagrid to Datagrids fromn akgulebubekir/Maui.DataGrid - Opening Page is quite faster and sorting in built.
