+++ FlowHub.Main/AppShell.xaml
@@ -29,11 +29,6 @@
                 ContentTemplate="{DataTemplate viewsExpenditures:ManageExpendituresPageD}"
                 Route="ManageExpenditures"/>
         </Tab>
-        <Tab Title="Flow Outs Depreciated" >
-            <ShellContent Title="Manage Expenditures"                          
-                ContentTemplate="{DataTemplate viewsExpenditures:ManageExpendituresD}"
-                Route="ManageExpenditures"/>
-        </Tab>
 
         <Tab Title="Flow In">
             <ShellContent Title="Manage Incomes"
