+++ FlowHub.Main/Platforms/Android/NavigationsMethods/ManageExpendituresNavs.cs
@@ -4,22 +4,22 @@ using FlowHub.Main.Views.Mobile.Statistics;
 
 namespace FlowHub.Main.Platforms.NavigationMethods;
 
-public class ManageExpendituresNavs
+public static class ManageExpendituresNavs
 {
-    public async void FromManageExpToUpsertExpenditures(Dictionary<string, object> navParams)
+    public static async Task FromManageExpToUpsertExpenditures(Dictionary<string, object> navParams)
     {
         await Shell.Current.GoToAsync(nameof(UpSertExpenditurePageM),true, navParams);
     }
-    public async void FromUpsertExpToManageExpenditures(Dictionary<string, object> navParams)
+    public static async Task FromUpsertExpToManageExpenditures(Dictionary<string, object> navParams)
     {
         await Shell.Current.GoToAsync(nameof(ManageExpendituresM), true, navParams);
     }
 
-    public async void FromManageExpToSingleMonthStats(Dictionary<string, object> navParams)
+    public static async Task FromManageExpToSingleMonthStats(Dictionary<string, object> navParams)
     {
         await Shell.Current.GoToAsync(nameof(SingleMonthStatsPageM), true, navParams);
     }
-    public async void ReturnOnce()
+    public static async Task ReturnOnce()
     {
         await Shell.Current.GoToAsync("..", true);
     }
